Getting rid of SemanticCheckerHelper.
