+++ src/crow/mir/mir_pass/mir_pass.cpp
@@ -1,6 +1,3 @@
 #include "mir_pass.hpp"
 
-
-namespace mir::mir_pass {
-
-} // 
+namespace mir::mir_pass {} // namespace mir::mir_pass
