+++ src/crow/mir/mir_pass/mir_pass.hpp
@@ -1,10 +1,17 @@
-#ifndef MIR_PASS_HPP
-#define MIR_PASS_HPP
+#ifndef CROW_CROW_MIR_MIR_PASS_MIR_PASS_HPP
+#define CROW_CROW_MIR_MIR_PASS_MIR_PASS_HPP
 
 // Absolute Includes:
 #include "crow/mir/mir.hpp"
 
 namespace mir::mir_pass {
+// Structs:
+struct MirPassParams {
+  ModulePtr m_module;
+  // TODO: SymbolTable.
+};
+
+// Classes:
 class MirPass {
   protected:
   ModulePtr m_module;
@@ -15,6 +22,11 @@ class MirPass {
   public:
   MirPass() = default;
 
+  // TODO: Think about this.
+  virtual auto on_function() -> void = 0;
+  virtual auto on_block() -> void = 0;
+  virtual auto on_instruction() -> void = 0;
+
   virtual auto run(ModulePtr t_module) -> void = 0;
 
   virtual ~MirPass() = default;
@@ -22,4 +34,4 @@ class MirPass {
 
 } // namespace mir::mir_pass
 
-#endif // MIR_PASS_HPP
+#endif // CROW_CROW_MIR_MIR_PASS_MIR_PASS_HPP
